Disable producer idempotency until KAFKA-13668 is fixed. (#980)
For the idempotent producer, some errors are permanent, meaning the producer instance becomes unusable afterwards. Authorization errors are one of these. This behaviour disagrees with the Kafka REST model, which assumes the producer to be stateless across requests. Until we either adapt Kafka REST to the idempotent model, or authorization errors become non-fatal, we will disable idempotent producers in Kafka REST.