KREST-4687 avro consume still fails
Looking at the logs for this intermittent failure, the consumer attempts to consume messages 4 times, over a course of 15 seconds, but only gets one record.

There are no logs relating to the produce (these are only written at trace level for the produce client) so it’s not possible to tell if the produces all went through.Produce is acks=all, the consumer’s auto.offset.reset is the default, which is latest.I’ve added in :

trace level logging for the produce and consume clients so we can diagnose failures more easily

logging of which offset we are getting back on the consume, so that we can see if it’s first, last or something else.  As part of this tidies up a set of nested try/catches into a while loop in consumeMessage

auto.offset.reset can now be set, and in the majority of cases is set to earliest.  This means that a new consumer group will read from the beginning, rather than there being the potential of missing some messages.  It’s not clear which is the single message being consumed, so I’ve added in a guess as to one reason that the other three messages are not got.  The one case it isn’t set to earliest is the consume by timestamp test

In case it’s the produce going awry, I check that the record metadata that is returned by the produce has an offset, and try again if it doesn’t.