Add a Fake Kafka/SR test environment.
Also convert ProduceActionIntegrationTest to use the new environment.

Using a real Kafka/SR implementation in our tests has the benefit that we can actually do an end-to-end test from Kafka REST to Kafka/SR. In the other hand, because of the distributed nature of Kafka, the tests become non-deterministic. This causes flakiness, which oftentimes we have to resort to sleep/retry to "fix".

This PR adds a fake Kafka/SR environment to address the aforementioned issues. The main design points are:

 1. No disk.
 2. No networking.
 3. No async.

By implementing Kafka following the above requirements, we can ensure tests are fully deterministic and fast(!!).

Note: I was mid-way implementing an AdminClient, until I noticed I didn't have to. This might explain some of the design decisions. I deleted all the code that was not used anymore, but the way the code is structured still reflects the extras needed for implementing the Admin operations. Eventually we will need to implement Admin anyway, so I think leaving the structure as is, without the extra code, is a good compromise.