KREST-4825 Huge timeout in consume test
The consumeTimeout test is failing in the 7.1 CP tests because of huge gaps between API calls.

There is 0.5s between the POST to create the consumer group and the subscription and then over 2 seconds between the subscription create and the GET that does the consume

Before this change the instance expiry (how long the consumer group is kept for) was set to 1s, and so the consumer group was timing out before the GET happened, resulting in a 404.

This PR increases the instance timeout to 3.5 s, which unfortutately adds 2.5s to the consumeTimeout test run time, but won't impact the run time of any other tests.