KREST-4961 Cherry-pick Kafka REST global Produce v3 rate limiter into v0.3137.x (#998)
* KREST-3323 add global rate limiter for produce (#993)

At the moment produce rate limiting is done per clusterId.  This means that if there are lots of clusterIds (as there would be in  MT cluster) then we could swamp the jvm.

This PR adds in a global count and byte rate limiter that will prevent this

The config is set to the same as the per clusterID limiters, so there is no change in behaviour for existing deployments.

* KREST-3323 Fix CI fail

The CI tests that use the injected produce rate limiter were failing, because the injected object was not properly initialised

I've swapped to using a provider.  If you get the provider in the constructor, to get the rate limiter, you get a NPE, so you need to do the get in the using method

* KREST-4979 reverse order of produce global rate limiters (#997)

Testing showed that putting the global rate limiter first results in reduced CPU load.

Co-authored-by: ehumber <87994408+ehumber@users.noreply.github.com>