KREST-5385 Add error_code to produce responses
Before this change, produce responses did not include an http status code if they were successful, only for error responses

This meant that code generated to match the openapi spec can't easily differentiate between a success and a failure response.

Ideally we'd have added in a status_code for success and failure cases, but there is an existing ErrorResponse that is used by all existing endpoints, and it seems overkill (and probably not the right thing to do for consistency) to put in place custom error handling for produce only, with a different structure
The response for a successful produce now looks like

{"error_code":200,"cluster_id":"nHQfD6V7Q_K29guDHiLdpw","topic_name":"summit","partition_id":0,"offset":6,"timestamp":"2022-04-22T14:48:47.650Z","value":{"type":"JSON","size":30}}

The error code is at the beginning to match what we do for the error response.

I thought about having a null error code response, but I think an explicit value is clearer.