KREST-5637 Add count based byte metrics and move tracking earlier
In our billing one pager, we have agreement that we will count billing metrics after rate limiting, but before validation of the produce request body.

This PR moves the produce metrics collection point slightly earlier to match that decision.

Currently the bytes metric is an Avg() - so if you send 10 messages of 67, the metric records 67.

I've added a summed count metric to the produce bytes sensor, which is what ce-kafka uses to record bytes as described here https://confluentinc.atlassian.net/wiki/spaces/KREST/pages/2777684595/Billing+metrics#Counting-the-bytes

Waiting on confirmation that this is correct, but I'm fairly sure it is, so raising the PR now.  It's simple to revisit later to swap to eg a Windowed Count or something else if needed.