KREST-5732 Refactor to allow us to override producer metrics in ce-kafka-rest
We need the ability to set custom tags on produce metrics in ce-kafka-rest, that aren't applicable to metrics collected by just kafka-rest.

To do this we need to be able to override the ProducerMetrics class that is used.

The ProducerMetrics instance is created in the ProducerMetricsFactory provide() method in v3ResourcesModule.

This change adds a ProducerMetrics Interface. This can then be implemented in kafka-rest and ce-kafka-rest, with different implementations of the ProducerMetrics - with a new ProducerMetricsFactory provide() method added in a new CloudV3ResourcesModule in ce-kafka-rest

99% of the implementation of ProducerMetrics is identical. All we need at the moment is the ability to set tags on the metrics defined in ProduceMetrics.  So, I've put existing code in an abstract class and then the implementation in kafka-rest does very little in addition. (An alternative would be to make the AbstractProducerMetrics not abstract but not final either).

There are a number of updates to AbstractProducerMetrics compared to producer metrics. The summary is
 - the ImmutableMap of tags is now an empty HashMap
 - there is a new method to add entries to the HashMap
 - The creation of the metrics.metricInstance now sets the tags in the MetricsTemplate (without this Kafka barfs saying For 'request-rate', runtime-defined metric tags do not match the tags in the template. Runtime = [tenant] Template = [])
 - Some fields are no longer private

For an example implementation in ce-kafka-rest see https://github.com/confluentinc/ce-kafka-rest/compare/master...ehumber:KREST-5732-add-tenant-tag-to-produce-metrics?expand=1