KREST-5830: Create a lazy wrapper around MappingIterator for Produce Action.
JsonParser instantiation, which happens during MappingIterator deserialization, can throw on bad request. This can cause the request scope to not be disposed properly if a `@RequestScoped` resoure is injected into a filter before an async resource method is executed. For Kafka REST, this only happens for the Produce Action.

This PR wraps the MappingIterator, so that the JsonParser is only instantiated during the first hasNext/nextValue call.

The PR also adds a test that shows the aforementioned resource leak.