KREST-6139 Deflake produce rate limiter test
There was a timing window in the test that checks the rate limiter cache expires.

The test is supposed to
 - send 1 message (OK)
 - send 2nd message (429)
 - expire rate limit cache
 - send 3rd message within 1 second : would get a 429 if using the same rate limiter instance, instead should get OK because the rate limiter has been reset

However what we see sometimes is

 - send 1st message (OK)
 - expire cache
 - send second message - get OK, not 429

There is an overriden time on the rate limiter, but this can't be overriden on the cache as that's outside our scope of control. That cache expire is managed by a thread.sleep :(

And it looks like the send of the second message, despite having an overriden time of only 1ms delay, was actually taking longer than the thread.sleep to really happen (slow jenkins :( )

I managed to reproduce this quite regularly by adding in a pile of log.trace statements.

Ideally we would mock up the rate limiter, but that's quite a big change, so for now I've just made all the gaps bigger.  Hopefully this is enough for jenkins, if not we'll have to revisit and mock up the actual rate limiter.