KREST-5637 Use a Meter as recommended by observability and the right Metrics Object (#1023)
* KREST-5637 Use a Meter as recommended by observability and the right Metrics Object

Our metrics weren't showing up in the telemetry pipeline, so obeservability suggested we use the same pattern for creating Metrics as the MultiTenantInterceptor does, which is a Meter.

I've basically copied the code over.

Even with this change the metrics didn't get added to the telemetry pipeline.  This was becaues each time a produce request arrived, we made a new Metrics object.  These are only registered with Telemetry Reporter on startup, and so the new metrics were never seen.

We should be using a single shared Metrics object for the whole of kafka-rest, not making a new one each time we add a new producer (this matches Kafka, which uses a single Metrics for all the server code).

The Metrics is definied in rest-utils, so I grab it from the Application and add it to the config object (which is where we already store the metrics tags) and pass it to ProducerMetrics from there.

* KREST-5637 address review comment