KREST-2655 Simple version of disconnect (#1039)
* KREST-2655 Simple version of disconnect

During our release review last year, one of the requests was to ensure that any long running connections are disconnected periodically.

Kafka does this for non-rest clients in https://github.com/confluentinc/ce-kafka/pull/3252 (which wasn't merged but has some interesting comments) and https://github.com/confluentinc/ce-kafka/pull/3292

The intent of this PR is to disconnect all streaming produce connections 24 hours after they were first made.  Once the stream has reached this age, the next message to arrive will get a 408 request timeout with message "Streaming connection open for longer than allowed: Connection will be closed."

Suggestions please for a better http status code, there is nothing that I could find that matches what we are doing.

We want to try and respond to any outstanding requests that we know about, rather than just losing them, so there is a grace period of 500ms, where any message already with us, or that arrive within this time period are also responded to with the 408.

Any messages that arrive after the grace period, but before the connection closing has finished up (a small timing window) are ignored.

The disconnect code only triggers for arriving messages, because there is already an existing, shorter timeout of 30s, for unused streams : https://github.com/confluentinc/rest-utils/blob/111998c26588cfcb773ee7bf9a757e3f79c5f301/core/src/main/java/io/confluent/rest/RestConfig.java#L418 so there was no point complicating the code further here to account for this.

hasNext() on the MappingIterator hangs if there is nothing new on the MappingIterator (you'd hope it would just return false)

There is still a timing window around "closingStarted" but it doesn't matter (i think) because all it means is that the scheduledExecutor has triggered the close on the MappingIterator and the output stream.  If we try and write to the closed output stream it does't matter, nothing bad will happen.  I didn't want to add a lock.

The next() call at line 163 is to clear the message off the MappingIterator so we can process the next one (and return 408) if needed.

There is a horrible override of clock (where I could have just an Instant for the stream start time) so that I can actually unit test this reliably - you can't mock Instant easily.

FUP - Should we add disconnect jitter in a follow on PR?

* KREST-2655 Address review comments

* KREST-2655 update how testing clock is overridden

* KREST-2655 last few nits