KREST-7005 use a sensor object per request - but shared per LKC id
In standalone kafka-rest the metricsTags are empty.

In ce-kafka-rest we inject the LKC tenant id as a tag

This PR adds the tenant tag to the end of the sensor name, so that each LKC has its own sensor.

This level of granularity is OK (rather than per individual request) because billing is done per lkc.

We also don't want a sensor per request as that would be a LOT of sensors.

Each sensor expires after being left unused for 1 hour.  Streams are disconnected after 30s if they are unused, so 1 hour should be OK.

I've tested this by copying over the ce-kafka-rest code to the V3ResourceModule and getting it to make a new LKC tag every other request that comes in
 - each lkc sensor has two requests written to it, then the next sensor is made.
 - the original sensor is not updated when the later sensor is created