Revert replacement of poll(Duration) with poll(long)
#1035 replaced a usage of the deprecated KafkaConsumer#poll(long)
API with the KafkaConsumer#poll(Duration) API in the internals of
Kafka REST's Consume v2 API. There is a slight difference in
behavior between the two KafkaConsumer APIs related to blocking
during consumer rebalancing.
- See https://cwiki.apache.org/confluence/display/KAFKA/KIP-266%3A+Fix+consumer+indefinite+blocking+behavior#KIP266:Fixconsumerindefiniteblockingbehavior-Consumer#poll

Unfortunately although this seemed fairly reasonable, we are
seeing change in behavior which breaks some tests and can affect
users of the Kafka REST Consume v2 API.

Because of that we will be reverting back to the old poll(long)
API until we find a better alternative.