KREST-7275 update openapi changes for batch acl
This commit
 - updates the kafka-rest openapi.yaml to include the changes for batch acl create which are required by the UI

 - It ports the code from Tom's ce-kafka-rest PR into kafka-rest, making this a new publically available API.  This feels like the right thing to do, because it's a simple API (just a wrapper around the individual create API comment), it was odd to split to very linked APIs across two repos (and this made our lives much harder wrt eg the integration test and the openapi doc).

 - Add a null check in the single create ACL to avoid a NPE and the subsequent 500 error being returned to the user.  Unfortunately this typed the cyclomatic complexity of all the valid payload checks in the createAcl class over the limit, so I had to pull them out into a separate method.

 - Add a log.error of any stack traces that are part of the failed async processing future.  Without these it's really hard to debug what goes wrong if any of our async calls fail.  Note that this change does not alter what the end user sees, just what comes out in our logs.

For batch ACL, if one of the structures within the data array is malformend  (eg an enum value doesn't match) then none of the  ACLs are created.  However, if one of the structures in the data array contains a value that the broker needs to determine is invalid, then the other structures are still applied.  For example if the user id specified in the first item in the array of ACLs does not actually exist, then that ACL is not created, but the second ACL in the array of ACLs is created (assuming it is valid)