KREST-8518 Add dry-run/validate-only versions of some topic APIs (#1090)
This change introduces an additional "validate only" parameter
to the requests for two important topic APIs:
- Create Topic
- Batch Alter Topic Config

When this parameter, which is a boolean flag, is set, the
corresponding operation would be dry-run, i.e. its execution
would be validated as if its actually run, but there would
be no side effects from the operation. E.g. if run in
that mode, a Create Topic request won't actually create a
topic.

As this requires just using simple overloads of the
corresponding Admin APIs, this change mostly focuses on
introducing the new parameters in minimally disruptive way
in exchange for some verbosity and maybe even redundancy.
- E.g. none of the existing tests had to be changed, including
  the resource test, which expect specific manager calls, and
  the manager impl tests, which expect specific Admin Client
  calls.