Address a coincidence causing a build breakage
ce-kafka-rest happens to have a config manager that
has defined on its own a validateOnly version of
safeAlterConfigs, so introducing one in the base
config manager (with a final modifier) causes a
breakage.

For now let's avoid the breakage by renaming the base
method and avoiding the unwanted override, and
later we can consider whether it's better to remove
the ce-kafka-rest method and use the base method
instead.