Handle null payload in POST requests (#1359)
Even though we annotate the payload with @Valid annotation, we might still receive null payload, this will cause the endpoint to return 500 due to NPE of the payload, this PR fixes the issue by checking the nullability of the payload and return 422 if that is the case.

The fix in this PR is for v3 resources.